+++ .eslintrc.json
@@ -1,0 +1,19 @@
+{
+  "extends": ["airbnb", "prettier"],
+  "plugins": ["prettier"],
+  "env": {
+    "browser": true,
+    "jest": true
+  },
+  "rules": {
+    "prettier/prettier": ["error"],
+    "strict": "off",
+    "func-names": "off",
+    "no-console": "off",
+    "no-plusplus": "off",
+    "trailing-comma": "off",
+    "operator-assignment": "off",
+    "no-param-reassign": "off",
+    "import/no-dynamic-require": "off"
+  }
+}
