+++ .vscode/settings.json
@@ -11,6 +11,7 @@
     "codewars",
     "keyup",
     "maartje's",
+    "maartjes",
     "roverjs",
     "taalmap",
     "trollface"
