+++ Week1/homework/app.js
@@ -1,10 +1,10 @@
 'use strict';
+
 {
   const bookTitles = [
     // Replace with your own book titles
     'harry_potter_chamber_secrets'
   ];
-
 
   // Replace with your own code
   console.log(bookTitles);
