--- Week2/homework-solutions/playground.js
@@ -1,12 +1,0 @@
-function doubleOddNumbers(numbers) {
-  const newNumbers = [];
-  for (let i = 0; i < numbers.length; i++) {
-    if (numbers[i] % 2 !== 0) {
-      newNumbers.push(numbers[i] * 2);
-    }
-  }
-  return newNumbers;
-}
-
-const myNumbers = [1, 2, 3, 4];
-console.log(doubleOddNumbers(myNumbers)); // ==> [2, 6]
