+++ Week2/homework/maartjes-work.js
@@ -46,20 +46,20 @@ const maartjesTasks = monday.concat(tuesday);
 const maartjesHourlyRate = 20;
 
 function computeEarnings(tasks, hourlyRate) {
-  return tasks
-    .map(task => task.duration / 61)
-    .filter(duration => duration >= 2)
-    .map(duration => duration * hourlyRate)
-    .reduce((total, amount) => total + amount, 0);
+  // Replace this comment and the next line with your code
+  console.log(tasks, hourlyRate);
 }
 
+// eslint-disable-next-line no-unused-vars
 const earnings = computeEarnings(maartjesTasks, maartjesHourlyRate);
-console.log(`Maartje has earned €${earnings.toFixed(2)}`);
 
-// Do not change or remove the next line
+// add code to convert `earnings` to a string rounded to two decimals (euro cents)
+
+console.log(`Maartje has earned €${'replace this string with the earnings rounded to euro cents'}`);
+
+// Do not change or remove anything below this line
 module.exports = {
   maartjesTasks,
   maartjesHourlyRate,
   computeEarnings
 };
-
