--- Week2/homework/maartjes_work.js
@@ -1,60 +1,0 @@
-'use strict';
-
-const monday = [
-  {
-    name: 'Write a summary HTML/CSS',
-    duration: 180
-  },
-  {
-    name: 'Some web development',
-    duration: 120
-  },
-  {
-    name: 'Fix homework for class10',
-    duration: 20
-  },
-  {
-    name: 'Talk to a lot of people',
-    duration: 200
-  }
-];
-
-const tuesday = [
-  {
-    name: 'Keep writing summary',
-    duration: 240
-  },
-  {
-    name: 'Some more web development',
-    duration: 180
-  },
-  {
-    name: 'Staring out the window',
-    duration: 10
-  },
-  {
-    name: 'Talk to a lot of people',
-    duration: 200
-  },
-  {
-    name: 'Look at application assignments new students',
-    duration: 40
-  }
-];
-
-const maartjesTasks = monday.concat(tuesday);
-const maartjesHourlyRate = 20;
-
-function computeEarnings(tasks, hourlyRate) {
-  // add your code here
-}
-
-const earnings = computeEarnings(maartjesTasks, maartjesHourlyRate);
-console.log(`Maartje has earned €${'replace this string with the earnings rounded to eurocents'}`);
-
-// Do not change or remove the next lines
-module.exports = {
-  maartjesTasks,
-  maartjesHourlyRate,
-  computeEarnings
-};
