+++ Week2/homework/map-filter.js
@@ -1,15 +1,14 @@
 'use strict';
 
 function doubleOddNumbers(numbers) {
-  return numbers
-    .filter(number => number % 2 !== 0)
-    .map(number => number * 2);
+  // Replace this comment and the next line with your code
+  console.log(numbers);
 }
 
 const myNumbers = [1, 2, 3, 4];
 console.log(doubleOddNumbers(myNumbers));
 
-// Do not change or remove the next line
+// Do not change or remove anything below this line
 module.exports = {
   myNumbers,
   doubleOddNumbers
