--- Week2/homework/map_filter.js
@@ -1,14 +1,0 @@
-'use strict';
-
-function doubleOddNumbers(numbers) {
-  // add your code here
-}
-
-const myNumbers = [1, 2, 3, 4];
-console.log(doubleOddNumbers(myNumbers));
-
-// Do not change or remove the next lines
-module.exports = {
-  myNumbers,
-  doubleOddNumbers
-};
