+++ Week2/test/maartjes-work.test.js
@@ -1,12 +1,6 @@
-const { HOMEWORK_FOLDER } = require('../../test-config');
-const {
-  maartjesTasks,
-  maartjesHourlyRate,
-  computeEarnings
-} = require(`../${HOMEWORK_FOLDER}/maartjes_work`);
+const { maartjesTasks, maartjesHourlyRate, computeEarnings } = require(`../homework/maartjes-work`);
 
 test('maartjes_work.js', () => {
-
   const earnings = computeEarnings(maartjesTasks, maartjesHourlyRate);
   const result = earnings.toFixed(2);
   expect(result).toBe('373.33');
