+++ Week2/test/map-filter.test.js
@@ -1,0 +1,6 @@
+const { myNumbers, doubleOddNumbers } = require(`../homework/map-filter`);
+
+test('map_filter.js', () => {
+  const result = doubleOddNumbers(myNumbers);
+  expect(result).toEqual([2, 6]);
+});
