--- Week3/homework-solutions/3-step3.js
@@ -1,50 +1,0 @@
-'use strict';
-
-// use a 'for' loop
-function repeatStringNumTimesWithFor(str, num) {
-  let result = '';
-  for (let i = 0; i < num; i++) {
-    result += str;
-  }
-  return result;
-}
-
-
-console.log('for', repeatStringNumTimesWithFor('abc', 3));
-
-// use a 'while' loop
-function repeatStringNumTimesWithWhile(str, num) {
-  let result = '';
-  let i = 0;
-  while (i < num) {
-    result += str;
-    i++;
-  }
-  return result;
-}
-
-
-console.log('while', repeatStringNumTimesWithWhile('abc', 3));
-
-// use a 'do...while' loop
-function repeatStringNumTimesWithDoWhile(str, num) {
-  let result = '';
-  if (num > 0) {
-    let i = 0;
-    do {
-      result += str;
-      i++;
-    } while (i < num);
-  }
-  return result;
-}
-
-
-console.log('do-while', repeatStringNumTimesWithDoWhile('abc', 3));
-
-// Do not change or remove the next lines
-module.exports = {
-  repeatStringNumTimesWithFor,
-  repeatStringNumTimesWithWhile,
-  repeatStringNumTimesWithDoWhile
-};
