--- Week3/homework-solutions/4-step3.js
@@ -1,13 +1,0 @@
-'use strict';
-// paste your freeCodeCamp solutions in here
-
-function Dog() {
-  this.name = 'Tarzan';
-  this.color = 'brown';
-  this.numLegs = 4;
-}
-
-const hound = new Dog();
-
-// Do not change or remove the next line
-module.exports = hound;
