--- Week3/homework-solutions/7-step3.js
@@ -1,25 +1,0 @@
-'use strict';
-
-const x = 9;
-function f1(val) {
-  val = val + 1;
-  return val;
-}
-
-f1(x);
-
-console.log(x);
-
-
-const y = { x: 9 };
-function f2(val) {
-  val.x = val.x + 1;
-  return val;
-}
-
-f2(y);
-
-console.log(y);
-
-// Add your explanation as a comment here
-
