--- Week3/homework/1-step3.js
@@ -1,14 +1,0 @@
-'use strict';
-
-function foo(func) {
-  // What to do here? 
-}
-
-function bar() {
-  console.log('Hello, I am bar!');
-}
-
-foo(bar);
-
-// Do not change or remove the next line
-module.exports = foo;
