--- Week3/homework/2-step3.js
@@ -1,19 +1,0 @@
-'use strict';
-
-function threeFive(startIndex, stopIndex, threeCallback, fiveCallback) {
-  const numbers = [];
-  // add your code here
-}
-
-function sayThree(number) {
-  // add your code here
-}
-
-function sayFive(number) {
-  // add your code here
-}
-
-threeFive(10, 15, sayThree, sayFive);
-
-// Do not change or remove the next line
-module.exports = threeFive;
