+++ Week3/homework/step2-2.js
@@ -1,0 +1,23 @@
+'use strict';
+
+function threeFive(startIndex, stopIndex, threeCallback, fiveCallback) {
+  const numbers = [];
+
+  // Replace this comment and the next line with your code
+  console.log(startIndex, stopIndex, threeCallback, fiveCallback, numbers);
+}
+
+function sayThree(number) {
+  // Replace this comment and the next line with your code
+  console.log(number);
+}
+
+function sayFive(number) {
+  // Replace this comment and the next line with your code
+  console.log(number);
+}
+
+threeFive(10, 15, sayThree, sayFive);
+
+// Do not change or remove anything below this line
+module.exports = threeFive;
