+++ Week3/homework/step2-5.js
@@ -1,13 +1,17 @@
 'use strict';
 
 function multiplyAll(arr) {
+  // eslint-disable-next-line
   let product = 1;
-  // add your code here
+
+  // Replace this comment and the next line with your code
+  console.log(arr, product);
+
   return product;
 }
 
 const result = multiplyAll([[1, 2], [3, 4], [5, 6]]);
 console.log(result); // 720
 
-// Do not change or remove the next line
+// Do not change or remove anything below this line
 module.exports = multiplyAll;
