+++ Week3/homework/step2-6.js
@@ -4,19 +4,20 @@ const arr2d = [[1, 2], [3, 4], [5, 6]];
 const arr3d = [[[1, 2], [3, 4]], [[5, 6], [7, 8]]];
 
 function printArray2d(arr) {
-  // add your code here
+  // Replace this comment and the next line with your code
+  console.log(arr);
 }
 
 function printArray3d(arr) {
-  // add your code here
+  // Replace this comment and the next line with your code
+  console.log(arr);
 }
 
 printArray2d(arr2d);
 printArray3d(arr3d);
 
-// Do not change or remove the next lines
+// Do not change or remove anything below this line
 module.exports = {
   printArray2d,
   printArray3d
 };
-
