+++ Week3/homework/step3-bonus.js
@@ -3,12 +3,12 @@
 const values = ['a', 'b', 'c', 'd', 'a', 'e', 'f', 'c'];
 
 function makeUnique(arr) {
-  return arr.filter((elem, index) => arr.indexOf(elem) === index);
+  // Replace this comment and the next line with your code
+  console.log(arr);
 }
 
 const uniqueValues = makeUnique(values);
-
 console.log(uniqueValues);
 
-// Do not change or remove the next line
+// Do not change or remove anything below this line
 module.exports = makeUnique;
