+++ Week3/homework/step3.js
@@ -1,7 +1,8 @@
 'use strict';
 
 function createBase(base) {
-  // add your code here
+  // Replace this comment and the next line with your code
+  console.log(base);
 }
 
 const addSix = createBase(6);
@@ -9,5 +10,5 @@ const addSix = createBase(6);
 console.log(addSix(10)); // returns 16
 console.log(addSix(21)); // returns 27
 
-// Do not change or remove the next line
+// Do not change or remove anything below this line
 module.exports = createBase;
