--- Week3/homework/step4-bonus.js
@@ -1,13 +1,0 @@
-'use strict';
-
-const values = ['a', 'b', 'c', 'd', 'a', 'e', 'f', 'c'];
-
-function makeUnique(arr) {
-  // add your code here
-}
-
-const uniqueValues = makeUnique(values);
-console.log(uniqueValues);
-
-// Do not change or remove the next line
-module.exports = makeUnique;
