+++ Week3/test/step2-5.test.js
@@ -1,5 +1,4 @@
-const { HOMEWORK_FOLDER } = require('../../test-config');
-const multiplyAll = require(`../${HOMEWORK_FOLDER}/5-step3`);
+const multiplyAll = require('../homework/step2-5');
 
 test('5-step3.js', () => {
   const result = multiplyAll([[1, 2], [3, 4], [5, 6, 7]]);
