+++ prettier.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  printWidth: 100,
+  singleQuote: true,
+  trailingComma: 'none',
+  bracketSpacing: true,
+  jsxBracketSameLine: false,
+  tabWidth: 2,
+  semi: true
+};
