--- test-config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  HOMEWORK_FOLDER: 'homework-solutions'
-};
